sidebar: Handle parallel space and user profile apps

+++ sidebar/app/src/main/java/com/libremobileos/sidebar/service/ServiceViewModel.kt
@@ -5,12 +5,18 @@ import android.app.prediction.AppPredictionContext
 import android.app.prediction.AppPredictionManager
 import android.app.prediction.AppPredictor
 import android.app.prediction.AppTarget
+import android.content.BroadcastReceiver
 import android.content.Context
+import android.content.Intent
+import android.content.Intent.ACTION_PROFILE_AVAILABLE
+import android.content.Intent.ACTION_PROFILE_UNAVAILABLE
+import android.content.IntentFilter
 import android.content.pm.LauncherApps
 import android.content.pm.PackageManager
 import android.os.Handler
 import android.os.HandlerExecutor
 import android.os.UserHandle
+import android.os.UserManager
 import androidx.appcompat.content.res.AppCompatResources
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
@@ -18,10 +24,13 @@ import com.libremobileos.sidebar.R
 import com.libremobileos.sidebar.app.SidebarApplication
 import com.libremobileos.sidebar.bean.AppInfo
 import com.libremobileos.sidebar.room.DatabaseRepository
+import com.libremobileos.sidebar.room.SidebarAppsEntity
 import com.libremobileos.sidebar.utils.Logger
 import com.libremobileos.sidebar.utils.contains
+import com.libremobileos.sidebar.utils.getBadgedIcon
 import com.libremobileos.sidebar.utils.getInfo
 import com.libremobileos.sidebar.utils.isResizeableActivity
+import com.libremobileos.sidebar.utils.isSidebarUserAllowed
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.cancelChildren
@@ -46,8 +55,10 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
 
     private val predictedAppList = MutableStateFlow<List<AppInfo>>(emptyList())
 
+    private val appContext = application.applicationContext
     private val launcherApps = application.getSystemService(Context.LAUNCHER_APPS_SERVICE)!! as LauncherApps
     private val appPredictionManager = application.getSystemService(AppPredictionManager::class.java)
+    private val userManager = application.getSystemService(UserManager::class.java)!!
 
     private var appPredictor: AppPredictor? = null
     private val handlerExecutor = HandlerExecutor(Handler())
@@ -55,7 +66,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
 
     val allAppActivity = AppInfo(
         "",
-        AppCompatResources.getDrawable(application.applicationContext, R.drawable.ic_all)!!,
+        AppCompatResources.getDrawable(appContext, R.drawable.ic_all)!!,
         ALL_APP_PACKAGE,
         ALL_APP_ACTIVITY,
         0
@@ -86,7 +97,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
             user: UserHandle?,
             replacing: Boolean
         ) {
-
+            logger.d("onPackagesAvailable: ${packageNames.contentToString()}, $user")
         }
 
         override fun onPackagesUnavailable(
@@ -94,13 +105,13 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
             user: UserHandle?,
             replacing: Boolean
         ) {
-
+            logger.d("onPackagesUnavailable: ${packageNames.contentToString()}, $user")
         }
     }
 
     private val appPredictionCallback = object : AppPredictor.Callback {
         override fun onTargetsAvailable(targets: List<AppTarget>) {
-            logger.d("appPredictionCallback targets: $targets")
+            logger.d("appPredictionCallback targets: ${targets.size}")
             predictedAppList.value = targets
                 .take(MAX_PREDICTED_APPS)
                 .mapNotNull { target ->
@@ -115,7 +126,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
                         } else {
                             AppInfo(
                                 info.loadLabel(application.packageManager).toString(),
-                                info.loadIcon(application.packageManager),
+                                application.getBadgedIcon(info, target.user),
                                 info.packageName,
                                 component.className,
                                 userId
@@ -125,6 +136,35 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
                         logger.e("failed to add $target: ", e)
                     }.getOrNull()
                 }
+        }
+    }
+
+    private val userProfileReceiver = object : BroadcastReceiver() {
+        override fun onReceive(context: Context, intent: Intent) {
+            val user: UserHandle = intent.getParcelableExtra(Intent.EXTRA_USER) ?: return
+            val userId = user.identifier
+            logger.d("userProfileReceiver received ${intent.action} $user")
+            when (intent.action) {
+                ACTION_PROFILE_AVAILABLE -> {
+                    val sidebarApps = repository.getAllSidebarWithoutLiveData() ?: return
+                    val newApps = sidebarApps
+                        .filter { it.userId == userId }
+                        .mapNotNull { entity ->
+                            runCatching { entity.toAppInfo() }
+                                .onFailure { e ->
+                                    logger.e("failed to add entity $entity: $e" )
+                                }
+                                .getOrNull()
+                        }
+                    logger.d("userProfileReceiver added apps: $newApps")
+                    // add them at the top
+                    _sidebarAppList.value = newApps + _sidebarAppList.value
+                }
+                ACTION_PROFILE_UNAVAILABLE -> {
+                    _sidebarAppList.value = _sidebarAppList.value
+                        .filter { it.userId != userId }
+                }
+            }
         }
     }
 
@@ -144,6 +184,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
         initSidebarAppList()
         launcherApps.registerCallback(launcherAppsCallback)
         registerAppPredictionCallback()
+        registerUserProfileReceiver()
         callbacksRegistered = true
     }
 
@@ -152,6 +193,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
         logger.d("unregisterCallbacks")
         launcherApps.unregisterCallback(launcherAppsCallback)
         appPredictor?.unregisterPredictionUpdates(appPredictionCallback)
+        appContext.unregisterReceiver(userProfileReceiver)
         viewModelScope.coroutineContext.cancelChildren()
         callbacksRegistered = false
     }
@@ -167,7 +209,7 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
             return
         }
         appPredictor = appPredictionManager.createAppPredictionSession(
-            AppPredictionContext.Builder(application.applicationContext)
+            AppPredictionContext.Builder(appContext)
                 .setUiSurface("hotseat")
                 .setPredictedTargetCount(MAX_PREDICTED_APPS)
                 .build()
@@ -177,35 +219,39 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
         }
     }
 
+    private fun registerUserProfileReceiver() {
+        appContext.registerReceiverAsUser(
+            userProfileReceiver,
+            UserHandle.CURRENT,
+            IntentFilter().apply {
+                addAction(ACTION_PROFILE_AVAILABLE)
+                addAction(ACTION_PROFILE_UNAVAILABLE)
+            },
+            null,
+            null
+        )
+    }
+
     private fun initSidebarAppList() {
         viewModelScope.launch(Dispatchers.IO) {
             repository.getAllSidebarAppsByFlow()
                 .combine(predictedAppList) { sidebarApps, predictedApps ->
                     mutableListOf<AppInfo>().apply {
-                        logger.d("sidebarApps=$sidebarApps predictedApps=$predictedApps")
+                        logger.d("initSidebarAppList: sidebarApps=$sidebarApps predictedApps=$predictedApps")
+                        // first add the pinned apps
                         sidebarApps?.forEach { entity ->
+                            if (!userManager.isSidebarUserAllowed(entity.userId)) {
+                                logger.w("initSidebarAppList: userid not allowed: $entity")
+                                return@forEach
+                            }
                             runCatching {
-                                if (!application.isResizeableActivity(entity.packageName, entity.activityName)) {
-                                    throw Exception("activity is not resizeable")
-                                }
-                                val info = application.packageManager.getApplicationInfo(entity.packageName, PackageManager.GET_ACTIVITIES)
-                                if (!info.enabled) {
-                                    throw Exception("package is disabled.")
-                                }
-                                add(
-                                    AppInfo(
-                                        "${info.loadLabel(application.packageManager)}",
-                                        info.loadIcon(application.packageManager),
-                                        entity.packageName,
-                                        entity.activityName,
-                                        entity.userId
-                                    )
-                                )
+                                add(entity.toAppInfo())
                             }.onFailure { e ->
-                                logger.w("initSidebarAppList: removing $entity: ", e)
+                                logger.w("initSidebarAppList: removing $entity: $e")
                                 repository.deleteSidebarApp(entity.packageName, entity.activityName, entity.userId)
                             }
                         }
+                        // then the predicted apps
                         addAll(
                             predictedApps.filter { sidebarApps?.contains(it)?.not() ?: true }
                         )
@@ -216,6 +262,26 @@ class ServiceViewModel(private val application: Application): AndroidViewModel(a
                     _sidebarAppList.value = sidebarAppList.toList()
                 }
         }
+    }
+
+    private fun SidebarAppsEntity.toAppInfo(): AppInfo {
+        if (!application.isResizeableActivity(packageName, activityName)) {
+            throw Exception("activity is not resizeable")
+        }
+        val info = application.packageManager.getApplicationInfo(
+            packageName,
+            PackageManager.GET_ACTIVITIES
+        )
+        if (!info.enabled) {
+            throw Exception("package is disabled.")
+        }
+        return AppInfo(
+            info.loadLabel(application.packageManager).toString(),
+            application.getBadgedIcon(info, userId),
+            packageName,
+            activityName,
+            userId
+        )
     }
 }
